Revert "normalize pattern correctly to avoid slow path in matching (#6288)"
This reverts commit dec61ad7505f148fb3f677e1dac9671ebae91550.
