+++ crates/turbopack-tests/tests/snapshot/imports/order/output/crates_turbopack-tests_tests_snapshot_imports_order_input_index_46327f.js
@@ -1,0 +1,6 @@
+(globalThis.TURBOPACK = globalThis.TURBOPACK || []).push([
+    "output/crates_turbopack-tests_tests_snapshot_imports_order_input_index_46327f.js",
+    {},
+    {"otherChunks":["output/crates_turbopack-tests_tests_snapshot_imports_order_input_6e1003._.js"],"runtimeModuleIds":["[project]/crates/turbopack-tests/tests/snapshot/imports/order/input/index.js [test] (ecmascript)"]}
+]);
+// Dummy runtime
