--- crates/turbopack-tests/tests/snapshot/imports/order/output/crates_turbopack-tests_tests_snapshot_imports_order_input_index_6fef85.js
@@ -1,11 +1,0 @@
-(globalThis.TURBOPACK = globalThis.TURBOPACK || []).push([
-    "output/crates_turbopack-tests_tests_snapshot_imports_order_input_index_6fef85.js",
-    {},
-]);
-(globalThis.TURBOPACK_CHUNK_LISTS = globalThis.TURBOPACK_CHUNK_LISTS || []).push({
-  "path": "output/crates_turbopack-tests_tests_snapshot_imports_order_input_index_6fef85.js",
-  "chunks": [
-    "output/crates_turbopack-tests_tests_snapshot_imports_order_input_17c023._.js"
-  ],
-  "source": "entry"
-});
