--- crates/turbopack-tests/tests/snapshot/imports/order/output/crates_turbopack-tests_tests_snapshot_imports_order_input_index_929e21.js
@@ -1,6 +1,0 @@
-(globalThis.TURBOPACK = globalThis.TURBOPACK || []).push([
-    "output/crates_turbopack-tests_tests_snapshot_imports_order_input_index_929e21.js",
-    {},
-    {"otherChunks":["output/crates_turbopack-tests_tests_snapshot_imports_order_input_17c023._.js"],"runtimeModuleIds":["[project]/crates/turbopack-tests/tests/snapshot/imports/order/input/index.js [test] (ecmascript)"]}
-]);
-// Dummy runtime
