+++ crates/turbopack-tests/tests/snapshot/imports/order/output/crates_turbopack-tests_tests_snapshot_imports_order_input_index_d81d26.js
@@ -1,0 +1,11 @@
+(globalThis.TURBOPACK = globalThis.TURBOPACK || []).push([
+    "output/crates_turbopack-tests_tests_snapshot_imports_order_input_index_d81d26.js",
+    {},
+]);
+(globalThis.TURBOPACK_CHUNK_LISTS = globalThis.TURBOPACK_CHUNK_LISTS || []).push({
+  "path": "output/crates_turbopack-tests_tests_snapshot_imports_order_input_index_d81d26.js",
+  "chunks": [
+    "output/crates_turbopack-tests_tests_snapshot_imports_order_input_6e1003._.js"
+  ],
+  "source": "entry"
+});
